+++ tests/testthat/test_simmr_input_plots.R
@@ -1,4 +1,5 @@
-context("Plot input data")
+set.seed(123)
+co <- function(expr) capture.output(expr, file = "NUL")
 
 data("geese_data_day1")
 simmr_1 <- with(
@@ -15,9 +16,12 @@ simmr_1 <- with(
 )
 
 test_that("basic simmr_input plot", {
-  expect_s3_class(plot(simmr_1), "ggplot")
-  expect_s3_class(plot(simmr_1, colour = FALSE), "ggplot")
-  expect_s3_class(plot(simmr_1, tracers = c(2, 1)), "ggplot")
+  p <- plot(simmr_1)
+  expect_doppelganger("simmr_input", p, path = 'simmr-input-plots')
+  p <- plot(simmr_1, colour = FALSE)
+  expect_doppelganger("simmr_input_no_col", p, path = 'simmr-input-plots')
+  p <- plot(simmr_1, tracers = c(2, 1))
+  expect_doppelganger("simmr_input_rev_tracers", p, path = 'simmr-input-plots')
 })
 
 test_that("1D simmr plot", {
@@ -34,10 +38,10 @@ test_that("1D simmr plot", {
     )
   )
 
-
-  # Plot 3 times - first default d13C vs d15N
-  expect_s3_class(plot(simmr_2), "ggplot")
-  expect_s3_class(plot(simmr_2, colour = FALSE), "ggplot")
+  p <- plot(simmr_2)
+  expect_doppelganger("simmr_input_1obs", p, path = 'simmr-input-plots')
+  p <- plot(simmr_2, colour = FALSE)
+  expect_doppelganger("simmr_input_1obs_nocol", p, path = 'simmr-input-plots')
 })
 
 test_that("Multi-groups plot", {
@@ -56,8 +60,10 @@ test_that("Multi-groups plot", {
     )
   )
 
-  expect_s3_class(plot(simmr_4), "ggplot")
-  expect_s3_class(plot(simmr_4, group = 5), "ggplot")
+  p <- plot(simmr_4)
+  expect_doppelganger("simmr_input_groups", p, path = 'simmr-input-plots')
+  p <- plot(simmr_4, group = 5)
+  expect_doppelganger("simmr_input_groups_specified", p, path = 'simmr-input-plots')
 })
 
 test_that("Single iso plot", {
@@ -76,7 +82,10 @@ test_that("Single iso plot", {
     )
   )
 
-  expect_s3_class(plot(simmr_5), "ggplot")
-  expect_s3_class(plot(simmr_5, group = 5), "ggplot")
-  expect_s3_class(plot(simmr_5, group = 5, colour = FALSE), "ggplot")
+  p <- plot(simmr_5)
+  expect_doppelganger("simmr_input_iso1", p, path = 'simmr-input-plots')
+  p <- plot(simmr_5, group = 5)
+  expect_doppelganger("simmr_input_iso1_group5", p, path = 'simmr-input-plots')
+  p <- plot(simmr_5, group = 5, colour = FALSE)
+  expect_doppelganger("simmr_input_iso1_group5_nocol", p, path = 'simmr-input-plots')
 })