+++ tests/testthat/test_simmr_print_summary.R
@@ -1,6 +1,6 @@
-context("simmr print and summary")
-
-library(simmr)
+set.seed(123)
+library(vdiffr)
+co <- function(expr) capture.output(expr, file = "NUL")
 
 data(geese_data_day1)
 
@@ -18,14 +18,14 @@ simmr_1 <- with(
   )
 )
 # MCMC run
-simmr_1_out <- simmr_mcmc(simmr_1,
+co(simmr_1_out <- simmr_mcmc(simmr_1,
   mcmc_control = list(
     iter = 100,
     burn = 10,
     thin = 1,
     n.chain = 2
   )
-)
+))
 
 # Taken from the simmr_mcmc example
 test_that("print.simmr_input", {
@@ -58,14 +58,14 @@ simmr_groups <- with(
     group = groups
   )
 )
-simmr_groups_out <- simmr_mcmc(simmr_groups,
+co(simmr_groups_out <- simmr_mcmc(simmr_groups,
   mcmc_control = list(
     iter = 100,
     burn = 10,
     thin = 1,
     n.chain = 2
   )
-)
+))
 # Taken from the simmr_mcmc example
 test_that("print.simmr_input", {
   expect_output(print(simmr_groups))
@@ -83,7 +83,7 @@ test_that("summary.simmr_output", {
 test_that("summary.simmr_output groups ", {
   expect_output(summary(simmr_groups_out, type = "statistics", group = 1))
   expect_output(summary(simmr_groups_out, type = "statistics", group = c(1, 3)))
-  summ_1 <- summary(simmr_groups_out, type = "statistics", group = 1)
-  summ_2 <- summary(simmr_groups_out, type = "statistics", group = 2)
+  co(summ_1 <- summary(simmr_groups_out, type = "statistics", group = 1))
+  co(summ_2 <- summary(simmr_groups_out, type = "statistics", group = 2))
   expect_false(summ_1$statistics[[1]][1, 1] == summ_2$statistics[[1]][1, 1])
 })